workshop-webinar/workshops/July-15-2025/customer_support_rag/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sentence-transformers==2.2.2
scikit-learn==1.3.0
scipy==1.10.1
rank-bm25==0.2.2
huggingface_hub==0.16.4  # ✅ add this line
